requirements.txt
action-tutorials-py==0.9.4
ament-clang-format==0.9.7
ament-copyright==0.9.7
ament-cppcheck==0.9.7
ament-cpplint==0.9.7
ament-flake8==0.9.7
ament-index-python==1.1.0
ament-lint==0.9.7
ament-lint-cmake==0.9.7
ament-package==0.9.5
ament-pep257==0.9.7
ament-uncrustify==0.9.7
ament-xmllint==0.9.7
# Editable install with no version control (andante-camera-calibration==0.0.0)
-e /home/ken/ros2_ws/andante_ws/build/andante_camera_calibration
# Editable install with no version control (andante-camera-linetrace==0.0.0)
-e /home/ken/ros2_ws/andante_ws/build/andante_camera_linetrace
# Editable install with no version control (andante-gazebo==0.0.0)
-e /home/ken/ros2_ws/andante_ws/build/andante_gazebo
barcode==1.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
demo-nodes-py==0.9.4
distro==1.8.0
domain-coordinator==0.9.2
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
idna==3.4
img2feat==0.2.0
joblib==1.2.0
joy-teleop==0.3.0
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
Levenshtein==0.20.8
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.6.0.66
osrf-pycommon==0.1.11
packaging==22.0
Pillow==9.3.0
PyQRCode==1.2.1
python-Levenshtein==0.20.8
PyYAML==6.0
pyzbar==0.1.9
pyzmq==24.0.1
quality-of-service-demo-py==0.9.4
rapidfuzz==2.13.6
requests==2.28.1
ros2action==0.9.12
ros2bag==0.3.9
ros2cli==0.9.12
ros2component==0.9.12
ros2doctor==0.9.12
ros2interface==0.9.12
ros2launch==0.11.7
ros2lifecycle==0.9.12
ros2multicast==0.9.12
ros2node==0.9.12
ros2param==0.9.12
ros2pkg==0.9.12
ros2run==0.9.12
ros2service==0.9.12
ros2topic==0.9.12
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt-action==0.4.9
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.0.5
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.3.0
scikit-build==0.16.3
scikit-learn==1.2.0
scipy==1.9.3
sros2==0.9.5
teleop-twist-keyboard==2.3.2
threadpoolctl==3.1.0
topic-monitor==0.9.4
torch==1.13.0
torchvision==0.14.0
typing-extensions==4.4.0
urllib3==1.26.13
zmq==0.0.0
